+++ src/test/java/com/blazemeter/taurus/junit/demotests/JUnit4Test.java
@@ -1,0 +1,12 @@
+package com.blazemeter.taurus.junit.demotests;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+public class JUnit4Test {
+    @Test
+    public void testJUnit4Method() {
+        assertEquals("1", "1");
+    }
+}